% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rre2-package.R
\name{Rre2-package}
\alias{Rre2}
\alias{Rre2-package}
\title{A wrapper around re2 for fast regular expression matching.}
\description{
R package \bold{Rre2} wraps the \href{https://github.com/google/re2}{re2}
library, a fast regular expression matching library written in C++. It
exports a reference class, RegExpRE2, that has the same API used in the
LikeThis package.
}
\seealso{
\code{\link{RegExpRE2}}
}

